[close #12] Update Format for pulling Memory Use
The format in the logs changed to use l2met format. This commit updates the app to pull the correct `memory_total` value.

The units of `memory_total` also changed from bytes to to megabytes, we no longer need to convert.

Also the original conversion was slightly off, there are 1048576 bytes to a mb.

We might want to talk to @fabiokung to see how stable this metric is, if it's likely to change we could add some logic that checks the unit type and does different unit conversion based on type. cc/ @dpiddy